requirements.txt
fastapi
uvicorn
gunicorn
strawberry-graphql[fastapi]
sqlalchemy
pytest
httpx